pyproject.toml
[tool.poetry]
name = "applifting-trainee"
version = "0.1.0"
description = ""
authors = ["michalomegalul <dog560154@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
psycopg2-binary = "^2.9.9"
requests = "^2.32.2"
python-dotenv = "^1.0.1"
apscheduler = "^3.10.4"
gunicorn = "^22.0.0"

[tool.poetry.dev-dependencies]
pytest = "^8.2.1"
pytest-cov = "^4.0.0"
tox = "^4.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
test = ["pytest", "pytest-cov", "tox"]
